backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
alembic
asyncpg
psycopg2-binary
pydantic
python-dotenv
bcrypt
pydantic[email]
llama-stack==0.2.14
llama_stack_client==0.2.14
fire
httpx
vulture
black
isort
pre-commit
flake8
httpx
kubernetes
